refactor(network-activity-plugin): move "cn" util to utils directory
