+++ packages/network-activity-plugin/src/ui/components/Input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '../utils';
+import { cn } from '../utils/cn';
 
 const Input = React.forwardRef<HTMLInputElement, React.ComponentProps<'input'>>(
   ({ className, type, ...props }, ref) => {
