+++ packages/network-activity-plugin/src/ui/components/Tabs.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import * as TabsPrimitive from '@radix-ui/react-tabs';
 
-import { cn } from '../utils';
+import { cn } from '../utils/cn';
 
 const Tabs = TabsPrimitive.Root;
